pyproject.toml
[project]
name = "confluence-summarizer"
version = "0.1.0"
description = "Um sistema robusto para resumir, analisar e reescrever páginas do Confluence usando RAG e Agentes de IA."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "httpx",
    "chromadb",
    "openai",
    "tenacity",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "flake8",
    "pyright",
    "respx",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.pyright]
include = ["src", "tests"]
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false
